+++ cmd/translatesfx/translatesfx/translate.go
@@ -22,6 +22,7 @@ type saCfgInfo struct {
 	globalDims       map[interface{}]interface{}
 	saExtension      map[string]interface{}
 	configSources    map[interface{}]interface{}
+	writer           map[interface{}]interface{}
 	monitors         []interface{}
 	observers        []interface{}
 	metricsToExclude []interface{}
@@ -39,8 +40,9 @@ func saExpandedToCfgInfo(saExpanded map[interface{}]interface{}) saCfgInfo {
 		saExtension:      saExtension(saExpanded),
 		observers:        observers(saExpanded),
 		configSources:    configSources(saExpanded),
-		metricsToExclude: metricsToExclude(saExpanded),
-		metricsToInclude: metricsToInclude(saExpanded),
+		metricsToExclude: saToMetricsToExclude(saExpanded),
+		metricsToInclude: saToMetricsToInclude(saExpanded),
+		writer:           writer(saExpanded),
 	}
 }
 
@@ -108,11 +110,11 @@ func configSources(saExpanded map[interface{}]interface{}) map[interface{}]inter
 	return cs
 }
 
-func metricsToExclude(saExpanded map[interface{}]interface{}) []interface{} {
+func saToMetricsToExclude(saExpanded map[interface{}]interface{}) []interface{} {
 	return toSlice(saExpanded, "metricsToExclude")
 }
 
-func metricsToInclude(saExpanded map[interface{}]interface{}) []interface{} {
+func saToMetricsToInclude(saExpanded map[interface{}]interface{}) []interface{} {
 	return toSlice(saExpanded, "metricsToInclude")
 }
 
@@ -126,4 +128,16 @@ func toSlice(saExpanded map[interface{}]interface{}, key string) []interface{} {
 		return nil
 	}
 	return l
+}
+
+func writer(saExpanded map[interface{}]interface{}) map[interface{}]interface{} {
+	v, ok := saExpanded["writer"]
+	if !ok {
+		return nil
+	}
+	wr, ok := v.(map[interface{}]interface{})
+	if !ok {
+		return nil
+	}
+	return wr
 }