+++ internal/configprovider/manager_test.go
@@ -64,7 +64,7 @@ func TestConfigSourceManager_NewManager(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			filename := path.Join("testdata", tt.file+".yaml")
-			parser, err := configparser.NewParserFromFile(filename)
+			parser, err := configparser.NewConfigMapFromFile(filename)
 			require.NoError(t, err)
 
 			manager, err := NewManager(parser, zap.NewNop(), component.DefaultBuildInfo(), tt.factories)
@@ -102,7 +102,7 @@ func TestConfigSourceManager_Simple(t *testing.T) {
 		},
 	}
 
-	cp := configparser.NewParserFromStringMap(originalCfg)
+	cp := configparser.NewConfigMapFromStringMap(originalCfg)
 
 	res, err := manager.Resolve(ctx, cp)
 	require.NoError(t, err)
@@ -135,7 +135,7 @@ func TestConfigSourceManager_ResolveRemoveConfigSourceSection(t *testing.T) {
 		"tstcfgsrc": &testConfigSource{},
 	})
 
-	res, err := manager.Resolve(context.Background(), configparser.NewParserFromStringMap(cfg))
+	res, err := manager.Resolve(context.Background(), configparser.NewConfigMapFromStringMap(cfg))
 	require.NoError(t, err)
 	require.NotNil(t, res)
 
@@ -170,26 +170,12 @@ func TestConfigSourceManager_ResolveErrors(t *testing.T) {
 				"tstcfgsrc": &testConfigSource{ErrOnRetrieve: testErr},
 			},
 		},
-		{
-			name: "error_on_retrieve_end",
-			config: map[string]interface{}{
-				"cfgsrc": "$tstcfgsrc:selector",
-			},
-			configSourceMap: map[string]configsource.ConfigSource{
-				"tstcfgsrc": &testConfigSource{
-					ErrOnRetrieveEnd: testErr,
-					ValueMap: map[string]valueEntry{
-						"selector": {Value: "test_value"},
-					},
-				},
-			},
-		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			manager := newManager(tt.configSourceMap)
 
-			res, err := manager.Resolve(ctx, configparser.NewParserFromStringMap(tt.config))
+			res, err := manager.Resolve(ctx, configparser.NewConfigMapFromStringMap(tt.config))
 			require.Error(t, err)
 			require.Nil(t, res)
 			require.NoError(t, manager.Close(ctx))
@@ -225,11 +211,11 @@ map:
 	})
 
 	file := path.Join("testdata", "yaml_injection.yaml")
-	cp, err := configparser.NewParserFromFile(file)
+	cp, err := configparser.NewConfigMapFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "yaml_injection_expected.yaml")
-	expectedParser, err := configparser.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewConfigMapFromFile(expectedFile)
 	require.NoError(t, err)
 	expectedCfg := expectedParser.ToStringMap()
 
@@ -254,11 +240,11 @@ func TestConfigSourceManager_ArraysAndMaps(t *testing.T) {
 	})
 
 	file := path.Join("testdata", "arrays_and_maps.yaml")
-	cp, err := configparser.NewParserFromFile(file)
+	cp, err := configparser.NewConfigMapFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "arrays_and_maps_expected.yaml")
-	expectedParser, err := configparser.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewConfigMapFromFile(expectedFile)
 	require.NoError(t, err)
 
 	res, err := manager.Resolve(ctx, cp)
@@ -304,11 +290,11 @@ func TestConfigSourceManager_ParamsHandling(t *testing.T) {
 	})
 
 	file := path.Join("testdata", "params_handling.yaml")
-	cp, err := configparser.NewParserFromFile(file)
+	cp, err := configparser.NewConfigMapFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "params_handling_expected.yaml")
-	expectedParser, err := configparser.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewConfigMapFromFile(expectedFile)
 	require.NoError(t, err)
 
 	res, err := manager.Resolve(ctx, cp)
@@ -340,7 +326,7 @@ func TestConfigSourceManager_WatchForUpdate(t *testing.T) {
 		},
 	}
 
-	cp := configparser.NewParserFromStringMap(originalCfg)
+	cp := configparser.NewConfigMapFromStringMap(originalCfg)
 	_, err := manager.Resolve(ctx, cp)
 	require.NoError(t, err)
 
@@ -394,7 +380,7 @@ func TestConfigSourceManager_MultipleWatchForUpdate(t *testing.T) {
 		},
 	}
 
-	cp := configparser.NewParserFromStringMap(originalCfg)
+	cp := configparser.NewConfigMapFromStringMap(originalCfg)
 	_, err := manager.Resolve(ctx, cp)
 	require.NoError(t, err)
 
@@ -443,11 +429,11 @@ func TestConfigSourceManager_EnvVarHandling(t *testing.T) {
 	})
 
 	file := path.Join("testdata", "envvar_cfgsrc_mix.yaml")
-	cp, err := configparser.NewParserFromFile(file)
+	cp, err := configparser.NewConfigMapFromFile(file)
 	require.NoError(t, err)
 
 	expectedFile := path.Join("testdata", "envvar_cfgsrc_mix_expected.yaml")
-	expectedParser, err := configparser.NewParserFromFile(expectedFile)
+	expectedParser, err := configparser.NewConfigMapFromFile(expectedFile)
 	require.NoError(t, err)
 
 	res, err := manager.Resolve(ctx, cp)