+++ internal/configprovider/parser.go
@@ -40,7 +40,7 @@ type (
 
 // Load reads the configuration for ConfigSource objects from the given parser and returns a map
 // from the full name of config sources to the respective ConfigSettings.
-func Load(ctx context.Context, v *configparser.Parser, factories Factories) (map[string]ConfigSettings, error) {
+func Load(ctx context.Context, v *configparser.ConfigMap, factories Factories) (map[string]ConfigSettings, error) {
 	processedParser, err := processParser(ctx, v)
 	if err != nil {
 		return nil, err
@@ -54,8 +54,8 @@ func Load(ctx context.Context, v *configparser.Parser, factories Factories) (map
 	return cfgSrcSettings, nil
 }
 
-// processParser prepares a configparser.Parser to be used to load config source settings.
-func processParser(ctx context.Context, v *configparser.Parser) (*configparser.Parser, error) {
+// processParser prepares a configparser.ConfigMap to be used to load config source settings.
+func processParser(ctx context.Context, v *configparser.ConfigMap) (*configparser.ConfigMap, error) {
 	// Use a manager to resolve environment variables with a syntax consistent with
 	// the config source usage.
 	manager := newManager(make(map[string]configsource.ConfigSource))
@@ -63,7 +63,7 @@ func processParser(ctx context.Context, v *configparser.Parser) (*configparser.P
 		_ = manager.Close(ctx)
 	}()
 
-	processedParser := configparser.NewParser()
+	processedParser := configparser.NewConfigMap()
 	for _, key := range v.AllKeys() {
 		if !strings.HasPrefix(key, configSourcesKey) {
 			// In Load we only care about config sources, ignore everything else.
@@ -86,7 +86,7 @@ func loadSettings(css map[string]interface{}, factories Factories) (map[string]C
 
 	// Iterate over extensions and create a config for each.
 	for key, value := range css {
-		settingsParser := configparser.NewParserFromStringMap(cast.ToStringMap(value))
+		settingsParser := configparser.NewConfigMapFromStringMap(cast.ToStringMap(value))
 
 		// Decode the key into type and fullName components.
 		componentID, err := config.NewIDFromString(key)