+++ internal/receiver/smartagentreceiver/factory_test.go
@@ -22,16 +22,15 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/config/configcheck"
+	"go.opentelemetry.io/collector/config/configtest"
 	"go.opentelemetry.io/collector/consumer/consumertest"
-	"go.uber.org/zap"
 )
 
 func TestCreateDefaultConfig(t *testing.T) {
 	factory := NewFactory()
 	cfg := factory.CreateDefaultConfig()
 	assert.NotNil(t, cfg, "failed to create default config")
-	assert.NoError(t, configcheck.ValidateConfig(cfg))
+	assert.NoError(t, configtest.CheckConfigStruct(cfg))
 }
 
 func TestCreateMetricsReceiver(t *testing.T) {
@@ -39,7 +38,7 @@ func TestCreateMetricsReceiver(t *testing.T) {
 	cfg := factory.CreateDefaultConfig()
 	cfg.(*Config).monitorConfig = &haproxy.Config{}
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	receiver, err := factory.CreateMetricsReceiver(context.Background(), params, cfg, consumertest.NewNop())
 	assert.NoError(t, err)
 	assert.NotNil(t, receiver)
@@ -52,7 +51,7 @@ func TestCreateMetricsReceiverWithInvalidConfig(t *testing.T) {
 	cfg := &Config{}
 	require.Error(t, cfg.validate())
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	receiver, err := factory.CreateMetricsReceiver(context.Background(), params, cfg, consumertest.NewNop())
 	require.Error(t, err)
 	assert.EqualError(t, err, "you must supply a valid Smart Agent Monitor config")
@@ -66,7 +65,7 @@ func TestCreateLogsReceiver(t *testing.T) {
 	cfg := factory.CreateDefaultConfig()
 	cfg.(*Config).monitorConfig = &haproxy.Config{}
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	receiver, err := factory.CreateLogsReceiver(context.Background(), params, cfg, consumertest.NewNop())
 	assert.NoError(t, err)
 	assert.NotNil(t, receiver)
@@ -79,7 +78,7 @@ func TestCreateLogsReceiverWithInvalidConfig(t *testing.T) {
 	cfg := &Config{}
 	require.Error(t, cfg.validate())
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	receiver, err := factory.CreateLogsReceiver(context.Background(), params, cfg, consumertest.NewNop())
 	require.Error(t, err)
 	assert.EqualError(t, err, "you must supply a valid Smart Agent Monitor config")
@@ -93,7 +92,7 @@ func TestCreateTracesReceiver(t *testing.T) {
 	cfg := factory.CreateDefaultConfig()
 	cfg.(*Config).monitorConfig = &haproxy.Config{}
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	receiver, err := factory.CreateTracesReceiver(context.Background(), params, cfg, consumertest.NewNop())
 	assert.NoError(t, err)
 	assert.NotNil(t, receiver)
@@ -106,7 +105,7 @@ func TestCreateTracesReceiverWithInvalidConfig(t *testing.T) {
 	cfg := &Config{}
 	require.Error(t, cfg.validate())
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	receiver, err := factory.CreateTracesReceiver(context.Background(), params, cfg, consumertest.NewNop())
 	require.Error(t, err)
 	assert.EqualError(t, err, "you must supply a valid Smart Agent Monitor config")
@@ -120,7 +119,7 @@ func TestCreateMetricsThenLogsAndThenTracesReceiver(t *testing.T) {
 	cfg := factory.CreateDefaultConfig()
 	cfg.(*Config).monitorConfig = &haproxy.Config{}
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	nextMetricsConsumer := consumertest.NewNop()
 	metricsReceiver, err := factory.CreateMetricsReceiver(context.Background(), params, cfg, nextMetricsConsumer)
 	assert.NoError(t, err)
@@ -150,7 +149,7 @@ func TestCreateTracesThenLogsAndThenMetricsReceiver(t *testing.T) {
 	cfg := factory.CreateDefaultConfig()
 	cfg.(*Config).monitorConfig = &haproxy.Config{}
 
-	params := component.ReceiverCreateSettings{Logger: zap.NewNop()}
+	params := component.ReceiverCreateSettings{}
 	nextTracesConsumer := consumertest.NewNop()
 	tracesReceiver, err := factory.CreateTracesReceiver(context.Background(), params, cfg, nextTracesConsumer)
 	assert.NoError(t, err)