+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 ## Unreleased
 
+## v0.36.0
+
+This Splunk OpenTelemetry Connector release includes changes from the [opentelemetry-collector v0.36.0](https://github.com/open-telemetry/opentelemetry-collector/releases/tag/v0.36.0) and the [opentelemetry-collector-contrib v0.36.0](https://github.com/open-telemetry/opentelemetry-collector-contrib/releases/tag/v0.36.0) releases.
+
+### 💡 Enhancements 💡
+
+- Handle backwards compatibility of internal ballast removal (#759)
+- Update bundled Smart Agent to [v5.14.1](https://github.com/signalfx/signalfx-agent/releases/tag/v5.14.1)
+- Automatically relocate removed OTLP exporter "insecure" field (#783)
+
+### 🧰 Bug fixes 🧰
+
+- Move Heroku buildpack to [https://github.com/signalfx/splunk-otel-collector-heroku](https://github.com/signalfx/splunk-otel-collector-heroku) (#755)
+- Fix rpm installation conflicts with the Smart Agent rpm (#773)
+
+## v0.35.0
+
+This Splunk OpenTelemetry Connector release includes changes from the [opentelemetry-collector v0.35.0](https://github.com/open-telemetry/opentelemetry-collector/releases/tag/v0.35.0) and the [opentelemetry-collector-contrib v0.35.0](https://github.com/open-telemetry/opentelemetry-collector-contrib/releases/tag/v0.35.0) releases.
+
+## 🚀 New components 🚀
+
+- [`groupbyattrs` processor](https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/processor/groupbyattrsprocessor)
+
+### 💡 Enhancements 💡
+
+- Update bundled Smart Agent to [v5.13.0](https://github.com/signalfx/signalfx-agent/releases/tag/v5.13.0) (#738)
+- Add SUSE support to [Linux installer script](https://github.com/signalfx/splunk-otel-collector/blob/main/docs/getting-started/linux-installer.md) (collector only, log collection with Fluentd not yet supported) (#720)
+- Add SUSE support to [puppet module](https://forge.puppet.com/modules/signalfx/splunk_otel_collector) (collector only, log collection with Fluentd not yet supported) (#737)
+
+### 🧰 Bug fixes 🧰
+
+- `smartagent` receiver: Properly parse receiver creator endpoints (#718)
+
 ## v0.34.1
 
 ### 💡 Enhancements 💡